[proliferate] move plans cache ownership
